Angular2 Decorator Template Highlighting
Added simple support for Angular2 inline template highlighting. In short, if we find something that looks like this:

@Component({
  template: `
    <some><html><here></here></html></some>
  `
})

We activate text.html.basic highlighting on the contents of the multiline string.

It's possible that we could look specifically for @Component, but right now, we're piggybacking off of #decorator to prevent weirdness/code duplication.

To make this work, I had to remove the $base include from #decorator. If people begin passing anonymous functions and so on to decorators, this will need revisiting. But right now, we're talking about booleans and objects, so I don't know that it really needs $base. If it does, probably better to just catch @Component->template on its own.