Allow using baseUrl, paths and rootDirs options in tsconfig,json (#996)
* allow baseUrl, paths, rootDirs in tsconfig files

Add the above as valid tsconfig.json properties.

Note: this change does not necessarily make these properties functional, it
simply allows them to be passed to the transpiler.

`baseUrl` and `paths` seem to work. `rootDirs` has not been verified and
might not be functioning (at least on cur. typescript@next).

TODO: this does not yet deeply validate the paths and rootDirs properties.

* Upgrade ntypescript to 1.201607050909.1

* Convert baseUrl and rootDirs options to absolute paths

* Fix a few type errors introduced by ntypescript upgrade

* Remove outdated comment